Fixed with "Kiselsteine"
Since there is no recursion we have to replace the backward mechanism with something similar.

Note that, in general, recursion uses the stack to remember the previous states. So it will also work building a stack
with the forward steps that can always be travel backwards.

This is implemented with the stack called Kiselsteine and the methods forward and backward.

With this solution the destination is reached also in the big board.